+++ src/commands/Jr.cc
@@ -5,7 +5,7 @@
 using namespace std;
 
 void Jr::run(MipsClient & client, const LabelTable & table) const {
-    int32_t s = client.getRegister(tokens.at(1).valAsNumber());
+    int64_t s = client.getRegister(tokens.at(1).valAsNumber());
     if (s == -1) throw EndProgram("Successfully ended program.");
     if (s % 4 != 0) throw InvalidScan("Unaligned Access to Program Counter");
     client.setPC(s);
